// api/generated/model/registerPushInstallationRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export interface RegisterPushInstallationRequest {
  /** @nullable */
  installationId?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  fcmToken?: string | null;
  /** @nullable */
  appVersion?: string | null;
  /** @nullable */
  environment?: string | null;
  /** @nullable */
  permissionStatus?: string | null;
}